-Renamed CaseTypeField to CaseTypeFilter - Moved CaseTypeFilter from reports.fields to reports.filters.select - updated references of reports.fields.CaseTypeField to reports.filters.select.CaseTypeFilter
